added and connected buttonand link to profile edit from main profile page
