+++ src/App.js
@@ -1,10 +1,12 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
 
+
 import Home from "./pages/Home";
 import Profile from "./pages/Profile";
 import NotFound from "./pages/NotFound";
-// import EditProfile from "./pages/EditProfile";
+
+
 
 class App extends React.Component {
   render() {
@@ -15,8 +17,9 @@ class App extends React.Component {
         <Route exact path="/" component={Home} />
         <Route exact path="/profile/:username" component={Profile} />
         <Route exact path="*" component={NotFound} />
-        {/* <Route exact path="/EditProfile/:edit-name" /> */}
-        {/* <EditProfile /> */}
+        
+        
+        
       </Switch>
       
     );
