+++ src/components/loginForm/LoginForm.js
@@ -2,6 +2,7 @@ import React from "react";
 import Spinner from "react-spinkit";
 import { withAsyncAction } from "../../redux/HOCs";
 import "./LoginForm.css";
+import { Input, Button } from 'semantic-ui-react'
 
 class LoginForm extends React.Component {
   constructor(props) {
@@ -27,7 +28,7 @@ class LoginForm extends React.Component {
       <div className="LoginForm">
         <form id="login-form" onSubmit={this.handleLogin}>
           <label htmlFor="username">Username</label>
-          <input
+          <Input
             type="text"
             name="username"
             autoFocus
@@ -35,15 +36,16 @@ class LoginForm extends React.Component {
             onChange={this.handleChange}
           />
           <label htmlFor="password">Password</label>
-          <input
+          <Input
             type="password"
             name="password"
             required
             onChange={this.handleChange}
           />
-          <button type="submit" disabled={loading}>
+          <br/>
+          <Button type="submit" size='large' disabled={loading}>
             Login
-          </button>
+          </Button>
         </form>
         {loading && <Spinner name="circle" color="blue" />}
         {error && <p style={{ color: "red" }}>{error.message}</p>}
