+++ src/components/menu/Menu.js
@@ -1,26 +1,41 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import "./Menu.css";
+
 import { withAsyncAction } from "../../redux/HOCs";
+import { logout } from "../../redux/stateReducers/auth";
 
 class Menu extends React.Component {
+    constructor(props){
+      super(props)
+
+      
+    }
+
+
   handleLogout = (event) => {
     event.preventDefault();
     this.props.logout();
   };
 
   render() {
+
+    
+    
     return (
       <div className="Menu">
         <h1>Something</h1>
+
+
         {this.props.isAuthenticated && (
           <div id="menu-links">
             <Link to="/pages/Profile.js">My Profile</Link>
-            <Link to="/pages/Profile.js">Photos</Link>
+            
             <Link to="/pages/Message.js">Message Feed</Link>
             <Link to="./" onClick={this.handleLogout}>
               Logout
             </Link>
+            
           </div>
         )}
       </div>
@@ -33,4 +48,5 @@ export default withAsyncAction("auth", "logout")(Menu);
             
               
            
+
