--- src/pages/EditProfile.js
@@ -1,25 +1,0 @@
-// import React from "react";
-// import { Link } from "react-router-dom";
-
-// // const { render } = require("react-dom");
-
-// render () {
-//     return
-// }
-// import React, { Component } from "react";
-// import { render } from "react-dom";
-
-// class EditProfile extends React.Component {
-//   render() {
-//     return (
-//       <div className="EditProfile">
-//         <form>
-//           <label htmlFor="EditProfile">Edit Profile</label>
-//           <input className="edit-name" type="input" name="displayname" />
-//           <button types="submit">Edit Profile</button>
-//         </form>
-//       </div>
-//     );
-//   }
-// }
-// export default EditProfile;
