+++ src/pages/ProfilePageMain.js
@@ -1,0 +1,47 @@
+import React from "react";
+
+import { userIsAuthenticated } from "../redux/HOCs";
+
+import { withRouter } from 'react-router-dom'
+import { Input, Button  } from 'semantic-ui-react'
+
+
+
+
+
+
+class ProfilePageMain extends React.Component {
+  constructor(props) {
+    super(props);
+
+    
+    
+  }
+
+  
+  
+  
+  
+  
+
+
+
+
+
+  
+  render() {
+    
+    return (
+      
+      <>
+      
+      </>
+      
+    )
+    
+  }
+  
+}
+
+
+export default userIsAuthenticated(ProfilePageMain);
